+++ Mensagem_R.xaml.cs
@@ -66,16 +66,11 @@ namespace JobStack
             Borda.Width = largura;
             Msg.Height = altura;
             Msg.Width = largura;
-
-
-
-
-
         }
 
-        public int getAltura()
+        public double getAltura()
         {
-            return altura + 25;
+            return altura+25;
         }
     }
 }
